Restrict environment keys for ImageMath.eval()
Prevent environment keys matching the names of builtins and keys
containing double underscores from being passed to ImageMath.eval(),
by raising a ValueError.

This is a cherry-pick of python-pillow/Pillow#7655 to fix
CVE-2023-50447 in Pillow-SIMD. Pillow only fixed it in 10.x, but
Pillow-SIMD doesn't have that release yet so it's vulnerable. This
commit solves that issue.